--- nest/src/routes/notes.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { NotesController } from './notes.controller';
-
-describe('NotesController', () => {
-  let controller: NotesController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [NotesController],
-    }).compile();
-
-    controller = module.get<NotesController>(NotesController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
