--- src/app.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { NotesModule } from './notes/notes.module';
-
-@Module({
-  imports: [NotesModule],
-})
-export class AppModule {}
