--- src/notes/notes.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('notes')
-export class NotesController {}
