--- src/notes/notes.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { NotesController } from './notes.controller';
-import { NotesService } from './notes.service';
-
-@Module({
-  controllers: [NotesController],
-  providers: [NotesService]
-})
-export class NotesModule {}
